+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Allow GraphQL tracing over other protocols in addition to HTTP.
+
 ## 1.84.1
 - Also capture incoming HTTP calls that time out on the client side or are aborted on the server side (via `req.destroy()`).